DjangoAPI/requirements.txt
asgiref==3.11.1
attrs==25.4.0
boto3==1.42.41
botocore==1.42.41
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==4.2.28
django-cors-headers==4.9.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
environ==1.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jmespath==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.11.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
stripe==14.3.0
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==1.26.20
whitenoise==6.11.0
